refactor: 디렉토리 구조 변경
- Entity 별로 정리
