+++ src/main/java/com/sedin/qna/article/model/Article.java
@@ -1,7 +1,7 @@
 package com.sedin.qna.article.model;
 
 import com.sedin.qna.account.model.Account;
-import com.sedin.qna.article.comment.model.Comment;
+import com.sedin.qna.comment.model.Comment;
 import com.sedin.qna.common.model.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.Builder;
