+++ src/main/java/com/sedin/qna/comment/service/CommentService.java
@@ -1,6 +1,6 @@
-package com.sedin.qna.article.comment.service;
+package com.sedin.qna.comment.service;
 
-import com.sedin.qna.article.comment.model.CommentDto;
+import com.sedin.qna.comment.model.CommentDto;
 
 import java.util.List;
