+++ src/test/java/com/sedin/qna/comment/model/CommentTest.java
@@ -1,5 +1,6 @@
-package com.sedin.qna.article.comment.model;
+package com.sedin.qna.comment.model;
 
+import com.sedin.qna.comment.model.Comment;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
